fix(naming): types, naming and enums
