+++ packages/graph-core/src/index.ts
@@ -1,66 +1,26 @@
 import { KnowledgeState } from "@bb/types";
-import type { KnowledgeDoc } from "@bb/types";
-import type { FileAnalysis } from "@bb/db-core";
+import type {
+  KnowledgeDoc,
+  NodeScope,
+  RepoSummaryPayload,
+  UpsertRepoNodeInput,
+  FolderSummaryPayload,
+  UpsertFolderNodeInput,
+  SnapshotFilesInput,
+  UpsertFileNodeInput,
+  GraphPingResult,
+} from "@bb/types";
 
-export interface NodeScope {
-  orgId: string;
-  knowledgeId: string;
-  repoId: string;
-}
-
-export interface RepoSummaryPayload {
-  purpose: string;
-  summary: string;
-  keywords: string[];
-  architecture: string;
-  dataFlow: string;
-  majorSubsystems: string[];
-  keyPatterns: string[];
-}
-
-export interface UpsertRepoNodeInput {
-  scope: NodeScope;
-  repoUrl: string;
-  branch: string;
-  summary: RepoSummaryPayload;
-}
-
-export interface FolderSummaryPayload {
-  purpose: string;
-  summary: string;
-  keywords: string[];
-  classes: string[];
-  functions: string[];
-  importsInternal: string[];
-  importsExternal: string[];
-  dependencyGraph: string;
-}
-
-export interface UpsertFolderNodeInput {
-  scope: NodeScope;
-  folderPath: string;
-  summary: FolderSummaryPayload;
-}
-
-export interface SnapshotFilesInput {
-  knowledgeId: string;
-  commitHash: string;
-}
-
-export interface UpsertFileNodeInput {
-  orgId?: string;
-  knowledgeId: string;
-  repoId?: string;
-  relativePath: string;
-  language: string;
-  sha: string;
-  sizeBytes: number;
-  analysis: FileAnalysis;
-  folderPath?: string;
-  isBigFile?: boolean;
-  totalChunks?: number;
-  totalTokenCount?: number;
-}
+export type {
+  NodeScope,
+  RepoSummaryPayload,
+  UpsertRepoNodeInput,
+  FolderSummaryPayload,
+  UpsertFolderNodeInput,
+  SnapshotFilesInput,
+  UpsertFileNodeInput,
+  GraphPingResult,
+};
 
 export interface IGraphKnowledgeRepository {
   upsertKnowledgeNode(doc: KnowledgeDoc): Promise<void>;
@@ -86,11 +46,6 @@ export interface IGraphRepoRepository {
 export interface IGraphIndexRepository {
   ensureKnowledgeIndexes(): Promise<void>;
   ensureFlatFolderIndexes(): Promise<void>;
-}
-
-export interface GraphPingResult {
-  ok: boolean;
-  latencyMs: number;
 }
 
 export interface IGraphDatabaseProvider {
