+++ packages/ingest-github/src/pipeline/pull-helpers.ts
@@ -1,11 +1,11 @@
 import { KnowledgeState } from "@bb/types";
-import { knowledge } from "@bb/db";
-import { knowledge as graphKnowledge } from "@bb/graph-db";
+import { knowledgeDb } from "@bb/db";
+import { knowledgeGraph } from "@bb/graph-db";
 import type { PipelineSummary } from "#src/types/pipeline.ts";
 
 export async function transitionState(knowledgeId: string, state: KnowledgeState): Promise<void> {
-  await knowledge.setKnowledgeState(knowledgeId, state);
-  await graphKnowledge.setKnowledgeStateInGraph(knowledgeId, state).catch(() => undefined);
+  await knowledgeDb.setKnowledgeState(knowledgeId, state);
+  await knowledgeGraph.setKnowledgeStateInGraph(knowledgeId, state).catch(() => undefined);
 }
 
 export function emptyPullSummary(commitHash: string): PipelineSummary {
