+++ packages/queue/src/github-pull.ts
@@ -1,5 +1,5 @@
 import { JobPriority, JobType, KnowledgeState, type GithubPullPayload } from "@bb/types";
-import { knowledge as dbKnowledge } from "@bb/db";
+import { knowledgeDb } from "@bb/db";
 import { _getQueue } from "./manager.ts";
 import { buildJobMessage, dedupeKey, mapPriority } from "./envelope.ts";
 import type { EnqueueOptions } from "./github-index.ts";
@@ -9,7 +9,7 @@ export async function enqueueGithubPull(payload: GithubPullPayload, opts: Enqueu
   const message = buildJobMessage(JobType.GithubPull, priority, payload);
   const jobId = dedupeKey(JobType.GithubPull, payload.knowledgeId);
 
-  await dbKnowledge.setKnowledgeState(payload.knowledgeId, KnowledgeState.Queued);
+  await knowledgeDb.setKnowledgeState(payload.knowledgeId, KnowledgeState.Queued);
 
   const queue = _getQueue(JobType.GithubPull);
   await queue.add(JobType.GithubPull, message, {
