+++ packages/types/src/database.ts
@@ -1,0 +1,10 @@
+export interface DeleteKnowledgeResult {
+  knowledgeDeleted: number;
+  rawDeleted: number;
+  statsDeleted?: number;
+}
+
+export interface DbPingResult {
+  ok: boolean;
+  latencyMs: number;
+}
