+++ packages/types/src/index.ts
@@ -1,4 +1,4 @@
-export { Config } from "./config.ts";
+export { Config, DbProviderType, GraphProviderType } from "./config.ts";
 export { JobType, JobPriority } from "./job.ts";
 export type {
   GithubIndexPayload,
@@ -18,6 +18,20 @@ export type {
   KnowledgeInfo,
   KnowledgeSource,
   LocalKnowledgeSource,
+  KnowledgeListEntry,
+  CommitHashRecord,
 } from "./knowledge.ts";
 export type { StatsCommitEntry, StatsRepoEntry, StatsResponse, StatsTotals } from "./stats.ts";
 export type { UsageDoc, ActivityDoc, UsageIncrement, ActivityInput } from "./usage.ts";
+export type { FileAnalysisSection, FileAnalysis, RawFileDoc } from "./analysis.ts";
+export type { DeleteKnowledgeResult, DbPingResult } from "./database.ts";
+export type {
+  NodeScope,
+  RepoSummaryPayload,
+  UpsertRepoNodeInput,
+  FolderSummaryPayload,
+  UpsertFolderNodeInput,
+  SnapshotFilesInput,
+  UpsertFileNodeInput,
+  GraphPingResult,
+} from "./graph.ts";
